MetadataBuilder API cleanup

+++ src/System.Reflection.Metadata/src/System/Reflection/BlobWriterImpl.cs
@@ -41,7 +41,7 @@ namespace Roslyn.Reflection
             throw new ArgumentOutOfRangeException("value");
         }
 
-        internal static void WriteCompressedInteger(ref BlobWriter writer, uint value)
+        internal static void WriteCompressedInteger(ref BlobWriter writer, int value)
         {
             unchecked
             {
@@ -55,7 +55,7 @@ namespace Roslyn.Reflection
                 }
                 else if (value <= MaxCompressedIntegerValue)
                 {
-                    writer.WriteUInt32BE(0xc0000000 | value);
+                    writer.WriteUInt32BE(0xc0000000 | (uint)value);
                 }
                 else
                 {
@@ -64,7 +64,7 @@ namespace Roslyn.Reflection
             }
         }
 
-        internal static void WriteCompressedInteger(BlobBuilder writer, uint value)
+        internal static void WriteCompressedInteger(BlobBuilder writer, int value)
         {
             unchecked
             {
@@ -78,7 +78,7 @@ namespace Roslyn.Reflection
                 }
                 else if (value <= MaxCompressedIntegerValue)
                 {
-                    writer.WriteUInt32BE(0xc0000000 | value);
+                    writer.WriteUInt32BE(0xc0000000 | (uint)value);
                 }
                 else
                 {
